Added a brk instruction.
The problem with the jump to scope end instruction was that it was
conditional which caused problems when we had definitely unconditional
scope breaks, because the next label didn't know that the code before
it, thus the stack contents as well were not reachable.
Now to jump to the end of the scope we use the brk instruction, which is
unconditional and the compiler always resets the result stack after it.