Attempted foreign poitners in VM, but decided not to proceed.
They were difficult to work with and would require a lot of work in VM
assembly, so we will just cheat and have the VM be aware of CZ, so that
we can use C to do it.
In a proper non bootstrap backend this would obviously be implemented
in the machine code directly.