Implemented structure types and reference support.
Now we are able to define custom structure types and have the inputs,
outputs, results and entries in structures be structures.
To be able to work with structures we also now have support for
references. They allow symbolic manipulation of types without having to
load their whole onto the stack. They work with both variables and
inputs of any type.